+++ node/log.go
@@ -10,17 +10,25 @@ import (
 	"time"
 )
 
+// logTimeFormat is the time format used for logging.
+const logTimeFormat = "2006-01-02 15:04:05.000000"
+
 // LogEntry represents one log entry.
 type LogEntry struct {
-	Time   time.Time // the time of the LogEntry, per the node's clock
-	NodeID string    // the ID of the node that created the LogEntry
-	Tag    string    // tags the LogEntry for categorization
-	Text   string    // the text message for the LogEntry
+	Time   time.Time // the time the entry was logged, per the node's clock
+	NodeID NodeID    // the ID of the node that created the entry
+	Tag    string    // tags the entry for categorization
+	Text   string    // the entry's text
 }
 
 // init registers LogEntry with the gob encoder
 func init() {
 	gob.Register(LogEntry{})
+}
+
+// GetLogEntry implements antler.LogEntry
+func (l LogEntry) GetLogEntry() LogEntry {
+	return l
 }
 
 // flags implements message
@@ -33,12 +41,11 @@ func (l LogEntry) handle(node *node) {
 	node.parent.Send(l)
 }
 
-// String returns the entry for display.
-func (e LogEntry) String() string {
-	t := e.Text
+func (l LogEntry) String() string {
+	t := l.Text
 	if strings.Contains(t, "\n") {
 		t = "⏎\n" + t
 	}
-	return fmt.Sprintf("%s %s %s: %s", e.Time.Format(readableTimeFormat),
-		e.NodeID, e.Tag, t)
+	return fmt.Sprintf("%s %s %s: %s", l.Time.Format(logTimeFormat),
+		l.NodeID, l.Tag, t)
 }