+++ node/time.go
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0
-// Copyright 2022 Pete Heist
-
-package node
-
-// readableTimeFormat is the time format used for human consumption.
-const readableTimeFormat = "2006-01-02 15:04:05.000000"